+++ medclin/src/main/java/br/com/medclin/business/PacienteBusiness.java
@@ -3,6 +3,8 @@
  */
 package br.com.medclin.business;
 
+import java.math.BigInteger;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.domain.Page;
@@ -46,7 +48,7 @@ public class PacienteBusiness implements IPacienteBusiness {
 	}
 
 	@Override
-	public Paciente buscarPacientePorCodigo(final Long codigoPaciente) {
+	public Paciente buscarPacientePorCodigo(final BigInteger codigoPaciente) {
 		return cloneUtil.clonePaciente(pacienteRep.buscarPacientePorCodigo(codigoPaciente));
 	}
 
@@ -68,7 +70,7 @@ public class PacienteBusiness implements IPacienteBusiness {
 	}
 
 	@Override
-	public void deletarPaciente(final Long codigoPaciente) {
+	public void deletarPaciente(final BigInteger codigoPaciente) {
 		pacienteRep.deleteById(codigoPaciente);
 	}
