+++ medclin/src/main/java/br/com/medclin/business/interfaces/IMedicoBusiness.java
@@ -1,5 +1,7 @@
 package br.com.medclin.business.interfaces;
 
+import java.math.BigInteger;
+
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 
@@ -11,11 +13,11 @@ public interface IMedicoBusiness {
 
 	public Page<Medico> buscarMedicoPorNome(final PageRequest pageable, final String nomeMedico);
 
-	public Medico buscarMedicoPorCodigo(final Long codigoMedico);
+	public Medico buscarMedicoPorCodigo(final BigInteger codigoMedico);
 
 	public Medico criarMedico(final Medico medico);
 
-	public void deletarMedico(final Long codigoMedico);
+	public void deletarMedico(final BigInteger codigoMedico);
 
 	public Page<Medico> listarMedico(final PageRequest pageable);
