+++ medclin/src/main/java/br/com/medclin/repository/MedicoRepository.java
@@ -3,6 +3,7 @@
  */
 package br.com.medclin.repository;
 
+import java.math.BigInteger;
 import java.util.Date;
 
 import org.springframework.data.domain.Page;
@@ -14,15 +15,15 @@ import org.springframework.data.repository.query.Param;
 
 import br.com.medclin.model.Medico;
 
-public interface MedicoRepository extends JpaRepository<Medico, Long>, PagingAndSortingRepository<Medico, Long> {
+public interface MedicoRepository extends JpaRepository<Medico, BigInteger>, PagingAndSortingRepository<Medico, BigInteger> {
 
 	@Query(value = "SELECT m FROM Medico m WHERE m.codigoPessoa = :codigoMedico")
-	Medico buscarMedicoPorCodigo(@Param("codigoMedico") final Long codigoMedico);
+	Medico buscarMedicoPorCodigo(@Param("codigoMedico") final BigInteger codigoMedico);
 
 	@Query(value = "SELECT m FROM Medico m WHERE m.nomePessoa LIKE %:nomeMedico%")
 	Page<Medico> buscarMedicoPorNome(@Param("nomeMedico") final String nomeMedico, final Pageable page);
 
 	@Query(value = "SELECT m.dataCriacao FROM Medico m WHERE m.codigoPessoa = :codigoMedico")
-	Date buscarDataCriacaoPorCodigo(@Param("codigoMedico") final Long codigoMedico);
+	Date buscarDataCriacaoPorCodigo(@Param("codigoMedico") final BigInteger codigoMedico);
 
 }
