+++ medclin/src/main/java/br/com/medclin/repository/PacienteRepository.java
@@ -3,6 +3,7 @@
  */
 package br.com.medclin.repository;
 
+import java.math.BigInteger;
 import java.util.Date;
 
 import org.springframework.data.domain.Page;
@@ -15,16 +16,16 @@ import org.springframework.data.repository.query.Param;
 import br.com.medclin.model.Paciente;
 
 public interface PacienteRepository
-		extends JpaRepository<Paciente, Long>, PagingAndSortingRepository<Paciente, Long> {
+		extends JpaRepository<Paciente, BigInteger>, PagingAndSortingRepository<Paciente, BigInteger> {
 
 	@Query(value = "SELECT p FROM Paciente p WHERE p.codigoPessoa = :codigoPaciente")
-	Paciente buscarPacientePorCodigo(@Param("codigoPaciente") final Long codigoPaciente);
+	Paciente buscarPacientePorCodigo(@Param("codigoPaciente") final BigInteger codigoPaciente);
 
 	@Query(value = "SELECT p FROM Paciente p WHERE p.nomePessoa LIKE %:nomePaciente%")
 	Page<Paciente> buscarPacientePorNome(@Param("nomePaciente") final String nomePaciente,
 			final Pageable page);
 
 	@Query(value = "SELECT p.dataCriacao FROM Paciente p WHERE p.codigoPessoa = :codigoPaciente")
-	Date buscarDataCriacaoPorCodigo(@Param("codigoPaciente") final Long codigoPaciente);
+	Date buscarDataCriacaoPorCodigo(@Param("codigoPaciente") final BigInteger codigoPaciente);
 
 }
